move FastPassMonitor to back rawls
